requirements.txt
# App web (app.py) + microserviço (inference_server.py) — um venv no desenvolvimento
Flask==2.0.1
Werkzeug==2.0.3
flask-cors>=4.0.0
requests>=2.28.0
numpy==1.26.4
pillow==10.2.0
ultralytics>=8.0.0
